kernels: Add config-test
This list dependencies for Landlock tests, especially the new 9p and
FULE filesystems tests.

It may not make sense to add these configuration options in the
kselftest config file because tests don't directly require them.  If the
filesystem type is not mounted on a specific path, the tests are just
skipped.

Without these kernel options set (e.g. Linux 5.15 without the filesystem
tests), the diod and bindfs commands would fail.

Fixes: 60363cb67974 ("guest: Mount a 9p filesystem")
Fixes: 2cbba83427e7 ("guest: Mount a FUSE filessytem")
Signed-off-by: Mickaël Salaün <mic@digikod.net>